delete CalculateStepMetricsUseCase and add StepMetricsRepository

+++ app/src/main/java/com/example/fit_sentinel/data/local/entity/DailyStepsEntity.kt
@@ -11,5 +11,5 @@ data class DailyStepsEntity(
     val lastUpdateTime: Long,
     val distanceKm: Double? = null,
     val caloriesBurned: Double? = null,
-    val estimatedTimeMinutes: Int? = null
+    val estimatedTimeMinutes: Long? = null
 )
